apalis_redis

Struct RedisStorage

pub struct RedisStorage<T, Conn = ConnectionManager, C = JsonCodec<Vec<u8>>> { /* private fields */ }

Represents a [Storage] that uses Redis for storage.

Implementations

impl<T: Serialize + DeserializeOwned, Conn> RedisStorage<T, Conn>

pub fn new(conn: Conn) -> Self

Start a new connection

pub fn new_with_config(conn: Conn, config: Config) -> Self

Start a connection with a custom config

pub fn new_with_codec<C>(conn: Conn, config: Config) -> RedisStorage<T, Conn, C>
where C: Codec + Sync + Send + 'static,

Start a new connection providing custom config and a codec

pub fn get_connection(&self) -> &Conn

Get current connection

pub fn get_config(&self) -> &Config

Get the config used by the storage

impl<T, Conn, C> RedisStorage<T, Conn, C>

pub fn get_codec(&self) -> &PhantomData<C>

Get the underlying codec details

impl<T, Conn, C> RedisStorage<T, Conn, C>
where Conn: ConnectionLike + Send + Sync + 'static, C: Codec<Compact = Vec<u8>> + Send + 'static,

pub async fn retry( &mut self, worker_id: &WorkerId, task_id: &TaskId, ) -> Result<i32, RedisError>
where T: Send + DeserializeOwned + Serialize + Unpin + Sync + 'static,

Attempt to retry a job

pub async fn kill( &mut self, worker_id: &WorkerId, task_id: &TaskId, ) -> Result<(), RedisError>
where T: Send + DeserializeOwned + Serialize + Unpin + Sync + 'static,

Attempt to kill a job

pub async fn enqueue_scheduled( &mut self, count: usize, ) -> Result<usize, RedisError>

Required to add scheduled jobs to the active set

pub async fn reenqueue_active( &mut self, job_ids: Vec<&TaskId>, ) -> Result<(), RedisError>

Re-enqueue some jobs that might be abandoned.

pub async fn reenqueue_orphaned( &mut self, count: i32, dead_since: DateTime<Utc>, ) -> Result<usize, RedisError>

Re-enqueue some jobs that might be orphaned after a number of seconds